tc: kill intelligently when prefilling
If we kill a TC object while prefilling, it could potentially misbehave:
after being killed, if it resumes, its table would be in an unknown
state.  This commit adds _finished_prefill, which is set to true only
when init_after_prefill is *complete*.  Therefore, if this is killed
before init_after_prefill finishes, it will be run again from scratch
when we restart.  For now, I'm forcing the actual *prefill* function
(which doesn't do so much work as init_after_prefill) to run to the end
every time.